requirements.txt
openpyxl

requests
bs4

plotly
ib_insync

jupyter
qgrid
jupyterlab
jupyterthemes
jupyter_contrib_nbextensions

aiokafka
redisworks


